Validation and error handling added for Athlete Handler.

+++ OdinAPI/handler/athlete.py
@@ -5,7 +5,17 @@ class AthleteHandler:
     
     def mapAthleteToDict(self,record):
         """
-        Converts an athlete record returned by the Athlete DAO into a dictionary.
+        Converts an athlete record returned by the Athlete DAO into a dictionary 
+        and returns it.
+
+        Creates a result dictionary an then maps the values from the record 
+        list into the dictionary before sending it to the caller.
+
+        Args:
+            record: An athlete record in the database with basic athlete information
+        Returns:
+            A dictionay with the basic athlete information in the record list given
+            mapped in it.
         """
         result = {}
         result['id'] = record[0]
@@ -19,10 +29,47 @@ class AthleteHandler:
         result['school'] = record[8]
         result['number'] = record[9]
         result['profilePicLink'] = record[10]
-        result['sportName'] = record[11]
-        record['aPosition'] = record[12]
-        record['aCategory'] = record[13]
+        result['sportName'] = record[11]        
         return result
+
+    def mapAthleteWithPositionsAndCategoriesToDict(self,records):
+        """
+        Converts athlete records returned by the Athlete DAO into a dictionary 
+        and returns it.This dictionary will contain the positions or categories
+        of an athlete if they have them.
+
+        Creates a result dictionary an then maps the values from the records 
+        list into the dictionary before sending it to the caller. This dictonary 
+        will contain positions or categories of an athlete if the records given have
+        a sublist.
+     
+
+        Args:
+            records: An athlete record list in the database with basic athlete information if
+                     if contains a sublist then it has positions or categories of an athlete.
+        Returns:
+            A dictionay with the basic athlete information in the records list given as well
+            as athlete positions and categories, if the records list has them, mapped in it.
+        """
+        try:
+            records[0][0]##Verfies if records has a sublist.
+        except:
+            return self.mapAthleteToDict(records)#If records does'nt have a sublist then the normal mapAthleteToDict function is called.
+        
+        result = self.mapAthleteToDict(records[0])#Get the athlete dictonary with the basic athlete information.
+        positions = {}
+        categories = {}
+        for record in records:
+            if record[12]:#Holds the position of the athlete if not null.
+                positions.update(dict(((record[12],record[13]),)))
+            if record[14]:#Holds the category of the athlete if not null.
+                categories.update(dict(((record[14],record[15]),)))
+        
+        result.update(dict((('athlete_positions',positions),)))
+        result.update(dict((('athlete_categories',categories),)))
+        return result
+
+                
     
     def getAthletesBySport(self,sID):
         """
@@ -40,14 +87,20 @@ class AthleteHandler:
             A JSON containing all valid athletes that participate in a
             specified sport by the given sport id.
         """
+        if not isinstance(sID,int):#Validate that the sport id is an integer.
+            return jsonify(Error = "The id of the sport must be an integer."),400
+
         dao = AthleteDAO()
-        result = dao.getAthletesBySport(sID)
-        if not result:
-            return jsonify(Error = "Athletes not found for the sport: {}.".format(sID)),404
-        mappedResult = []
-        for athlete in result:                        
-            mappedResult.append(self.mapAthleteToDict(athlete))
-        return jsonify(Atheletes = mappedResult), 200
+        try:            
+            result = dao.getAthletesBySport(sID)
+            if not result:
+                return jsonify(Error = "Athletes not found for the sport: {}.".format(sID)),404
+            mappedResult = []
+            for athlete in result:   
+                mappedResult.append(self.mapAthleteWithPositionsAndCategoriesToDict(athlete))
+            return jsonify(Atheletes = mappedResult), 200
+        except:
+            return jsonify(Error = "An error ocurred while trying to get the athletes of a sport."),400
     
     def getAthleteByID(self,aID):
         """
@@ -64,12 +117,20 @@ class AthleteHandler:
             A JSON containing the athlete with the given id.
 
         """
+        if not isinstance(aID,int):#Validate that the athlete id is an integer.
+            return jsonify(Error = "The id of the athlete must be an integer."),400
+
         dao = AthleteDAO()
-        result = dao.getAthleteByID(aID)
-        if not result:
-            return jsonify(Error = "Athlete with aID:{} not found.".format(aID)),404
-        mappedResult = self.mapAthleteToDict(result)
-        return jsonify(Athlete = mappedResult), 200
+        try:  
+            
+            result = dao.getAthleteByID(aID)
+            if not result:
+                return jsonify(Error = "Athlete with aID:{} not found.".format(aID)),404
+            
+            mappedResult = self.mapAthleteWithPositionsAndCategoriesToDict(result)
+            return jsonify(Athlete = mappedResult), 200
+        except:
+            return jsonify(Error = "An error ocurred while trying to get an athelte by id."),400
     
     #Don't think this function is needed anymore.
     def getAthleteByName(self,aFName,aMName,aLName):
@@ -98,23 +159,30 @@ class AthleteHandler:
             A JSON object containing the id of the newly added user.
 
         """
+        if not isinstance(sID,int):#Validate that the sport id is an integer.
+            return jsonify(Error =  "The sport id must be in integer."),400
+
+        validationResult = self._validateAttributesList(attributes)
+        if isinstance(validationResult,str):
+            return jsonify(Error = validationResult),400
+        
         dao = AthleteDAO()
-        result = None
-        if attributes[10] and not attributes[11]:#attributes[11] contains the positions, #attributes[11] contains the categories
-            if not isinstance(attributes[10],dict):
-                return jsonify(Error = 'Positions for the athlete must be a dictionary.'),500
-            result = dao.addAthleteWithPosition(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[10])
-        elif not attributes[10] and attributes[11]:
-            if not isinstance(attributes[11],dict):
-                return jsonify(Error = 'Categories for the athlete must be a dictionary.'),500
-            result = dao.addAthleteWithCategory(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[11])
-        else:
-            result = dao.addAthlete(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9])
-        if isinstance(result,str):#If true, result will contain the error message.
-            return jsonify(Error = result),500
+        try:            
+            result = None
+            if aPositions and not aCategories:                   
+                result = dao.addAthleteWithPosition(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[10])
+            elif not aPositions and aCategories:                 
+                result = dao.addAthleteWithCategory(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[11])
+            else:
+                result = dao.addAthlete(sID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9])
+            
+            if isinstance(result,str):#If true, result will contain the error message.
+                return jsonify(Error = result),400
 
-        return jsonify(Athlete = "Added new athlete with id:{}.".format(result)),201
-
+            return jsonify(Athlete = "Added new athlete with id:{}.".format(result)),201
+        except:
+            return jsonify(Error = "Problem ocurred when adding a new athlete."),400
+            
     def editAthlete(self,aID,attributes):
         """
         Edits the attributes of an existing and valid athlete record with
@@ -130,14 +198,22 @@ class AthleteHandler:
         
         Returns:
             A JSON object containing the information of the edited athlete.
-        """
-        
+        """ 
+        if not isinstance(aID,int):
+            return jsonify(Error =  "The athlete id must be in integer."),400
+
+        validationResult = self._validateAttributesList(attributes)
+        if isinstance(validationResult,str):
+            return jsonify(Error = validationResult),400               
         dao = AthleteDAO()
-        result = dao.editAthlete(aID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[10],attributes[11],attributes[12])
-        if not result:
-            return jsonify(Error = "Athlete not found with id:{}.".format(aID)),404
-        mappedResult = self.mapAthleteToDict(result)
-        return jsonify(Athlete = mappedResult),200
+        try:            
+            result = dao.editAthlete(aID,attributes[0],attributes[1],attributes[2],attributes[3],attributes[4],attributes[5],attributes[6],attributes[7],attributes[8],attributes[9],attributes[10],attributes[11])
+            if isinstance(result,str):
+                return jsonify(Error = result),400
+            
+            return jsonify(Athlete = "Edited athlete with id: {}".format(result)),200
+        except:
+            return jsonify(Error = "Problem ocurred when updating an athlete."),400
     
     def removeAthlete(self,aID):
         """
@@ -152,13 +228,98 @@ class AthleteHandler:
         Returns:
             A JSON containing the id of the invalidated athlete.
         """
+        if not isinstance(aID,int):
+            return jsonify(Error =  "The athlete id must be in integer."),400
+        try:
+            dao = AthleteDAO()
+            result = dao.removeAthlete(aID)
+            if not result:
+                return jsonify(Error = "Athlete not found with id:{}.".format(aID)),404
+            return jsonify(Athlete = "Removed athlete with id:{}.".format(result)),200
+        except:
+            return jsonify(Error = "Problem ocurred when removing an athlete.")
 
-        dao = AthleteDAO()
-        result = dao.removeAthlete(aID)
-        if not result:
-            return jsonify(Error = "Athlete not found with id:{}.".format(aID)),404
-        return jsonify(Athlete = "Removed athlete with id:{}.".format(result)),200
-    
+    def _validateAttributesList(self,attributes):
+        """
+        Validates the attributes list given for the addAthlete() and editAthlete()
+        functions.
+
+        Args:
+            attributes: A list containing the attributes of an athelete to be added or 
+                        edited.
+        Returns:
+            A string with an error message if the validation fails an integer otherwise.        
+        """
+
+        if not isinstance(attributes,list):
+            return "The attributes of the athlete must be in a list."
+        if not len(attributes) == 12:
+            return "The attributes list does not have the correct ammount of elements."
+        
+        
+        #Going to extract the inputs from the attributes list.
+        aFName = attributes[0]
+        aMName = attributes[1]
+        aLName = attributes[2]
+        aBio = attributes[3]
+        aHeight = attributes[4]
+        aStudyProgram = attributes[5]
+        aDateOfBirth = attributes[6]
+        aSchoolOfPrecedence = attributes[7]
+        aNumber = attributes[8]
+        aProfilePictureLink = attributes[9]
+        aPositions = attributes[10]
+        aCategories = attributes[11]    
+   
+        #Validation of inputs 
+        if not aFName or not isinstance(aFName,str) or len(aFName)<2 or len(aFName)>20 or not aFName.isalpha():
+            return "First name given does not follow the constraints."
+
+        if aMName:
+            if not isinstance(aMName,str) or len(aMName)<2 or len(aMName)>20 or not aMName.isalpha():
+                return "Middle name given does nto follow the constraints."
+
+        if not aLName or not isinstance(aLName,str) or len(aLName)<2 or len(aLName)>40:#Still need to validates all the characters in the last name string.
+            return "Last name given does nto follow the constraints."
+
+        if aBio:
+            if not aBio or not isinstance(aBio,str) or len(aBio)<2 or len(aBio)>1000:
+                return "Athlete bio given does nto follow the constraints."
+
+        if aHeight:
+            if not isinstance(aBio,float):
+                return "The height given is not a real number."
+
+        if aStudyProgram:
+            if not isinstance(aStudyProgram,str):
+                return "The study program given is not a string."
+
+        if aDateOfBirth:
+            if not isinstance(aDateOfBirth,str):
+                return "The date of birth given is not valid."
+
+        if aSchoolOfPrecedence:
+            if not isinstance(aSchoolOfPrecedence,str):
+                return "The input given for school of precedence is incorrect."
+
+        if aNumber:
+            if not isinstance(aNumber,int):
+                return "The number of the athlete must be an integer."
+
+        if aProfilePictureLink:
+            if not isinstance(aProfilePictureLink,str):
+                return "The profile picture link must be a string."
+
+        if aPositions:
+            if not isinstance(aPositions,dict):
+                return "Positions must be in dictionary."
+
+        if aCategories:
+            if not isinstance(aCategories,dict):
+                return "Categories must be in dictionary."  
+
+        return 1            
+
